refactor(core): inline show* macro impls in UniversalOps companion
- show* / showRaw* / showType* impls now in `object UniversalOps` (companion of
  the value-class extension), placed in same file next to their `inline def`
- sourceCode / withSourceCode impls remain in a separate file
  (`commons.macros.UniversalOpsSourceMacros`) because the package object's
  auto-imported `universalOps` implicit makes `Position.sourceCode` recursively
  resolve to `UniversalOps.sourceCode` if placed in the same package
